+++ dfirtrack_config/forms.py
@@ -1,7 +1,7 @@
 from django import forms
 from django.contrib.auth.models import User
 from dfirtrack_artifacts.models import Artifactstatus
-from dfirtrack_config.models import ArtifactExporterSpreadsheetXlsConfigModel, MainConfigModel, SystemExporterMarkdownConfigModel, SystemExporterSpreadsheetCsvConfigModel, SystemExporterSpreadsheetXlsConfigModel, SystemImporterFileCsvConfigbasedConfigModel, SystemImporterFileCsvConfigModel, SystemImporterFileCsvFormbasedConfigModel
+from dfirtrack_config.models import ArtifactExporterSpreadsheetXlsConfigModel, MainConfigModel, SystemExporterMarkdownConfigModel, SystemExporterSpreadsheetCsvConfigModel, SystemExporterSpreadsheetXlsConfigModel, SystemImporterFileCsvConfigModel
 from dfirtrack_main.models import Analysisstatus, Case, Company, Dnsname, Domain, Location, Os, Reason, Recommendation, Serviceprovider, Systemstatus, Systemtype, Tag
 import os
 
@@ -288,175 +288,6 @@ class SystemExporterSpreadsheetXlsConfigForm(forms.ModelForm):
             'spread_xls_worksheet_recommendation': 'Export worksheet to explain recommendation',
             'spread_xls_worksheet_tag': 'Export worksheet to explain tag',
         }
-
-class SystemImporterFileCsvConfigbasedConfigForm(forms.ModelForm):
-    """ system importer CSV config form (config based only) """
-
-    # reorder field choices
-    csv_default_reason = forms.ModelChoiceField(
-        label = 'Set reason',
-        queryset = Reason.objects.order_by('reason_name'),
-        required = False,
-    )
-
-    # reorder field choices
-    csv_default_domain = forms.ModelChoiceField(
-        label = 'Set domain',
-        queryset = Domain.objects.order_by('domain_name'),
-        required = False,
-    )
-
-    # reorder field choices
-    csv_default_dnsname = forms.ModelChoiceField(
-        label = 'Set DNS name',
-        queryset = Dnsname.objects.order_by('dnsname_name'),
-        required = False,
-    )
-
-    # reorder field choices
-    csv_default_systemtype = forms.ModelChoiceField(
-        label = 'Set systemtype',
-        queryset = Systemtype.objects.order_by('systemtype_name'),
-        required = False,
-    )
-
-    # reorder field choices
-    csv_default_os = forms.ModelChoiceField(
-        label = 'Set OS',
-        queryset = Os.objects.order_by('os_name'),
-        required = False,
-    )
-
-    # reorder field choices
-    csv_default_location = forms.ModelChoiceField(
-        label = 'Set location',
-        queryset = Location.objects.order_by('location_name'),
-        required = False,
-    )
-
-    # reorder field choices
-    csv_default_serviceprovider = forms.ModelChoiceField(
-        label = 'Set serviceprovider',
-        queryset = Serviceprovider.objects.order_by('serviceprovider_name'),
-        required = False,
-    )
-
-    # reorder field choices
-    csv_default_case = forms.ModelMultipleChoiceField(
-        label = 'Set cases',
-        queryset = Case.objects.order_by('case_name'),
-        required = False,
-        widget=forms.CheckboxSelectMultiple(),
-    )
-
-    # reorder field choices
-    csv_default_company = forms.ModelMultipleChoiceField(
-        label = 'Set companies',
-        queryset = Company.objects.order_by('company_name'),
-        required = False,
-        widget=forms.CheckboxSelectMultiple(),
-    )
-
-    # reorder field choices
-    csv_default_tag = forms.ModelMultipleChoiceField(
-        label = 'Set tags',
-        queryset = Tag.objects.order_by('tag_name'),
-        required = False,
-        widget=forms.CheckboxSelectMultiple(),
-    )
-
-    # reorder field choices
-    csv_default_systemstatus = forms.ModelChoiceField(
-        queryset = Systemstatus.objects.order_by('systemstatus_name'),
-        label = 'Set systemstatus (*)',
-        required = True,
-        widget = forms.RadioSelect(),
-    )
-
-    # reorder field choices
-    csv_default_analysisstatus = forms.ModelChoiceField(
-        queryset = Analysisstatus.objects.order_by('analysisstatus_name'),
-        label = 'Set analysisstatus (*)',
-        required = True,
-        widget = forms.RadioSelect(),
-    )
-
-    class Meta:
-
-        # model
-        model = SystemImporterFileCsvConfigbasedConfigModel
-
-        # this HTML forms are shown
-        fields = (
-            'csv_skip_existing_system',
-            'csv_column_system',
-            'csv_headline',
-            'csv_choice_ip',
-            'csv_remove_ip',
-            'csv_column_ip',
-            'csv_remove_case',
-            'csv_remove_company',
-            'csv_remove_tag',
-            'csv_default_systemstatus',
-            'csv_default_analysisstatus',
-            'csv_default_reason',
-            'csv_default_domain',
-            'csv_default_dnsname',
-            'csv_default_systemtype',
-            'csv_default_os',
-            'csv_default_location',
-            'csv_default_serviceprovider',
-            'csv_default_case',
-            'csv_default_company',
-            'csv_default_tag',
-        )
-
-        labels = {
-            'csv_skip_existing_system': 'Skip existing systems',
-            'csv_column_system': 'Number of the column in the CSV file that contains the system name',
-            'csv_headline': 'CSV file contains a headline row',
-            'csv_choice_ip': 'CSV file contains IP addresses',
-            'csv_remove_ip': 'Remove / overwrite existing IP addresses for already existing systems',
-            'csv_column_ip': 'Number of the column in the CSV file that contains the IP addresses',
-            'csv_remove_case': 'Remove / overwrite existing cases for already existing systems',
-            'csv_remove_company': 'Remove / overwrite existing companies for already existing systems',
-            'csv_remove_tag': 'Remove / overwrite existing tags for already existing systems',
-        }
-
-        widgets = {
-            'csv_column_system': forms.NumberInput(
-                attrs={
-                    #'style': 'width:6ch',
-                    'min': '1',
-                    'max': '99',
-                    'size': '3',
-                },
-            ),
-            'csv_column_ip': forms.NumberInput(
-                attrs={
-                    #'style': 'width:6ch',
-                    'min': '1',
-                    'max': '99',
-                    'size': '3',
-                },
-            ),
-        }
-
-    def clean(self):
-        """ custom field validation """
-
-        # get form data
-        cleaned_data = super().clean()
-
-        # get relevant values
-        csv_column_system = self.cleaned_data['csv_column_system']
-        csv_column_ip = self.cleaned_data['csv_column_ip']
-
-        # compare column values
-        if csv_column_system == csv_column_ip:
-            raise forms.ValidationError('The columns for system and IP must not have the same values.')
-
-        return cleaned_data
 
 class SystemImporterFileCsvConfigForm(forms.ModelForm):
     """ system importer CSV config form """
@@ -1093,73 +924,5 @@ class SystemImporterFileCsvConfigForm(forms.ModelForm):
         # finally raise validation error
         if validation_errors:
             raise forms.ValidationError(validation_errors)
-
-        return cleaned_data
-
-class SystemImporterFileCsvFormbasedConfigForm(forms.ModelForm):
-    """ system importer CSV config form (form based only) """
-
-    class Meta:
-
-        # model
-        model = SystemImporterFileCsvFormbasedConfigModel
-
-        # this HTML forms are shown
-        fields = (
-            'csv_skip_existing_system',
-            'csv_column_system',
-            'csv_headline',
-            'csv_choice_ip',
-            'csv_remove_ip',
-            'csv_column_ip',
-            'csv_remove_case',
-            'csv_remove_company',
-            'csv_remove_tag',
-        )
-
-        labels = {
-            'csv_skip_existing_system': 'Skip existing systems',
-            'csv_column_system': 'Number of the column in the CSV file that contains the system name',
-            'csv_headline': 'CSV file contains a headline row',
-            'csv_choice_ip': 'CSV file contains IP addresses',
-            'csv_remove_ip': 'Remove / overwrite existing IP addresses for already existing systems',
-            'csv_column_ip': 'Number of the column in the CSV file that contains the IP addresses',
-            'csv_remove_case': 'Remove / overwrite existing cases for already existing systems',
-            'csv_remove_company': 'Remove / overwrite existing companies for already existing systems',
-            'csv_remove_tag': 'Remove / overwrite existing tags for already existing systems',
-        }
-
-        widgets = {
-            'csv_column_system': forms.NumberInput(
-                attrs={
-                    #'style': 'width:6ch',
-                    'min': '1',
-                    'max': '99',
-                    'size': '3',
-                },
-            ),
-            'csv_column_ip': forms.NumberInput(
-                attrs={
-                    #'style': 'width:6ch',
-                    'min': '1',
-                    'max': '99',
-                    'size': '3',
-                },
-            ),
-        }
-
-    def clean(self):
-        """ custom field validation """
-
-        # get form data
-        cleaned_data = super().clean()
-
-        # get relevant values
-        csv_column_system = self.cleaned_data['csv_column_system']
-        csv_column_ip = self.cleaned_data['csv_column_ip']
-
-        # compare column values
-        if csv_column_system == csv_column_ip:
-            raise forms.ValidationError('The columns for system and IP must not have the same values.')
 
         return cleaned_data
