--- dfirtrack_config/templates/dfirtrack_config/system/system_importer_file_csv_form_based_config_popup.html
@@ -1,140 +1,0 @@
-<!-- extending maintemplate -->
-{% extends "dfirtrack_main/mainpopup.html" %}
-
-<!-- HTML head in maintemplate -->
-{% block head %}
-
-<!-- page title -->
-<title>
-    System CSV file import config (form-based)
-</title>
-
-{% load static %}
-
-<!-- HTML head in maintemplate -->
-{% endblock %}
-
-<!-- ############################################ -->
-
-<!-- HTML body in maintemplate -->
-{% block content %}
-
-<!-- headline -->
-<div class="row col bottom-distance">
-    <h4>
-        System CSV file import config (form-based)
-    </h4>
-</div>
-
-<!-- explanation text -->
-<div class="row">
-    <div class="col">
-        <div class="card card-body bg-dark border-light top-distance bottom-distance copy-false">
-        You can choose how different system attributes are handled during CSV file import.
-        All other settings will be set in the file upload form.
-        </div>
-    </div>
-</div>
-
-<!-- form -->
-<form method="POST" class="post-form">
-
-    {% csrf_token %}
-
-    <!-- show form related errors -->
-    {% include "dfirtrack_main/includes/form_errors.html" %}
-
-    <!-- general settings block -->
-    <div class="row">
-
-        <div class="col">
-            <div class="card card-body bg-dark border-light top-distance">
-
-                <h5>
-                    General settings
-                </h5>
-
-                <table>
-                    <tr><td>{{ form.csv_skip_existing_system }}</td> <td>{{ form.csv_skip_existing_system.label }}</td></tr>
-                    <tr><td>{{ form.csv_column_system }}</td> <td>{{ form.csv_column_system.label }}</td></tr>
-                    <tr><td>{{ form.csv_headline }}</td> <td>{{ form.csv_headline.label }}</td></tr>
-                </table>
-
-            </div>
-        </div>
-
-    <!-- general settings block -->
-    </div>
-
-    <!-- IP related settings block -->
-    <div class="row">
-
-        <div class="col">
-            <div class="card card-body bg-dark border-light top-distance">
-
-                <h5>
-                    IP related settings
-                </h5>
-
-                <table class="bottom-distance">
-                    <tr><td>{{ form.csv_choice_ip }}</td> <td>{{ form.csv_choice_ip.label }}</td></tr>
-                </table>
-
-                The following settings are only taken into account if IP addresses have been selected.
-
-                <table class="top-distance">
-                    <tr><td>{{ form.csv_remove_ip }}</td> <td>{{ form.csv_remove_ip.label }}</td></tr>
-                    <tr><td>{{ form.csv_column_ip }}</td> <td>{{ form.csv_column_ip.label }}</td></tr>
-                </table>
-
-            </div>
-        </div>
-
-    <!-- IP related settings block -->
-    </div>
-
-    <!-- overriding settings block -->
-    <div class="row">
-
-        <div class="col">
-            <div class="card card-body bg-dark border-light top-distance">
-
-                <h5>
-                    Overriding settings
-                </h5>
-
-                The following settings will remove existing many-to-many relationships, before adding new ones (only applies for already existing systems).
-
-                <table class="top-distance">
-                    <tr><td>{{ form.csv_remove_case }}</td> <td>{{ form.csv_remove_case.label }}</td></tr>
-                    <tr><td>{{ form.csv_remove_company }}</td> <td>{{ form.csv_remove_company.label }}</td></tr>
-                    <tr><td>{{ form.csv_remove_tag }}</td> <td>{{ form.csv_remove_tag.label }}</td></tr>
-                </table>
-
-            </div>
-        </div>
-
-    <!-- overriding settings block -->
-    </div>
-
-    <!-- button row -->
-    <div class="row top-distance">
-
-        <!-- save button -->
-        <div class="col-4">
-            {% include "dfirtrack_main/includes/button_save.html" %}
-        </div>
-
-        <!-- cancel button -->
-        <div class="col-1">
-            {% include "dfirtrack_main/includes/button_cancel_popup.html" %}
-        </div>
-
-    <!-- button row -->
-    </div>
-
-<!-- form -->
-</form>
-
-<!-- HTML body in maintemplate -->
-{% endblock %}
