+++ dfirtrack_config/urls.py
@@ -1,7 +1,7 @@
 from django.urls import path
 from dfirtrack_config.exporter.markdown import system_exporter_markdown_config_editor
 from dfirtrack_config.exporter.spreadsheet import artifact_exporter_spreadsheet_config_editor, system_exporter_spreadsheet_config_editor
-from dfirtrack_config.importer.file import csv_config_editor
+from dfirtrack_config.importer.file import system_importer_file_config_editor
 from dfirtrack_config.views import main_config_editor, status, statushistory
 
 urlpatterns = [
@@ -11,9 +11,7 @@ urlpatterns = [
     path(r'system/exporter/markdown/', system_exporter_markdown_config_editor.system_exporter_markdown_config_view, name='system_exporter_markdown_config_popup'),
     path(r'system/exporter/spreadsheet/csv/', system_exporter_spreadsheet_config_editor.system_exporter_spreadsheet_csv_config_view, name='system_exporter_spreadsheet_csv_config_popup'),
     path(r'system/exporter/spreadsheet/xls/', system_exporter_spreadsheet_config_editor.system_exporter_spreadsheet_xls_config_view, name='system_exporter_spreadsheet_xls_config_popup'),
-    path(r'system/importer/file/csv/configbased/', csv_config_editor.system_importer_file_csv_config_based_config_view, name='system_importer_file_csv_config_based_config_popup'),
-    path(r'system/importer/file/csv/formbased/', csv_config_editor.system_importer_file_csv_form_based_config_view, name='system_importer_file_csv_form_based_config_popup'),
-    path(r'system/importer/file/csv/', csv_config_editor.system_importer_file_csv_config_view, name='system_importer_file_csv_config_popup'),
+    path(r'system/importer/file/csv/', system_importer_file_config_editor.system_importer_file_csv_config_view, name='system_importer_file_csv_config_popup'),
 
     path(r'status/', status.StatusView.as_view(), name='status'),
     path(r'status/<int:pk>/', status.StatusDetailView.as_view(), name='status_detail'),
